added language selector and rtl compatibility
This PR adds a variety of features to the LinkSwap dApp.
It contains a language switch, implemented into the header menu.
The feature utilizes the react-i18next library which was already implemented.
Following translations were done by the community and added to the dapp:
- Arabic (ar)
-- The Arabic translation was done by the discord members 'BraveHeart#2338', 'Jazz#9482' and 'yfL God#9469'
- Chinese (cn)
-- The Chinese translation was done by the discord member 'saythink#0191'
- German (de)
-- The German translation was done by the discord member 'Linkie Link#1393'
- German (de)
-- The German translation was done by the discord member 'Linkie Link#1393'
- Spanish (es)
-- The Spanish translation was done by the discord member 'Markos#6029'
- Italian (it)
-- The Italian translation was done by 'Roy#1776' himself
- Korean (kor)
-- The Korean translation was done by the discord member 'CJ CHOI#1945'
- Dutch (nl)
-- The Dutch translation was done by the discord members 'Atlalok#6716' and 'ChainLinkTCO#4560'
- Romanian (ro)
-- The Romanian translation was done by the discord member 'inartin#9707'
- Russian (ru)
-- The Russian translation was done by the discord member 'inartin#9707'
The Language Select was added as new react component and has a subcomponent called LanguageOption. The LanguageOption component enables a modular approach to add new languages in the future.
With LinkSwap evolving new translation strings will be added to the locales.